+++ lib/router/login/login_widget.dart
@@ -382,7 +382,7 @@ class _LoginSignupState extends ConsumerState<LoginSignupWidget> {
     _doPreLogin();
 
     // Set up the private key and nostr client
-    settingsProvider.addAndChangePrivateKey(privateKey, updateUI: true);
+    settingsProvider.addAndChangePrivateKey(privateKey);
     nostr = await relayProvider.genNostrWithKey(privateKey);
     ref.invalidate(groupIdentifierRepositoryProvider);
 
@@ -461,7 +461,7 @@ class _LoginSignupState extends ConsumerState<LoginSignupWidget> {
       _doPreLogin();
 
       var npubKey = Nip19.encodePubKey(pubkey!);
-      settingsProvider.addAndChangePrivateKey(npubKey, updateUI: true);
+      settingsProvider.addAndChangePrivateKey(npubKey);
 
       var pubkeyOnlySigner = PubkeyOnlyNostrSigner(pubkey);
       nostr = await relayProvider.genNostr(pubkeyOnlySigner);
@@ -484,7 +484,7 @@ class _LoginSignupState extends ConsumerState<LoginSignupWidget> {
           bunkerLink =
               (nostr!.nostrSigner as NostrRemoteSigner).info.toString();
         }
-        settingsProvider.addAndChangePrivateKey(bunkerLink, updateUI: true);
+        settingsProvider.addAndChangePrivateKey(bunkerLink);
       } finally {
         cancel.call();
       }
@@ -504,7 +504,7 @@ class _LoginSignupState extends ConsumerState<LoginSignupWidget> {
 
       _doPreLogin();
 
-      settingsProvider.addAndChangePrivateKey(pk, updateUI: true);
+      settingsProvider.addAndChangePrivateKey(pk);
       nostr = await relayProvider.genNostrWithKey(pk);
     }
 
@@ -533,7 +533,7 @@ class _LoginSignupState extends ConsumerState<LoginSignupWidget> {
     if (StringUtil.isNotBlank(androidNostrSigner.getPackage())) {
       key = "$key?package=${androidNostrSigner.getPackage()}";
     }
-    settingsProvider.addAndChangePrivateKey(key, updateUI: true);
+    settingsProvider.addAndChangePrivateKey(key);
     nostr = await relayProvider.genNostr(androidNostrSigner);
 
     if (backAfterLogin && mounted) {
@@ -556,7 +556,7 @@ class _LoginSignupState extends ConsumerState<LoginSignupWidget> {
     _doPreLogin();
 
     var key = "${NIP07Signer.uriPre}:$pubkey";
-    settingsProvider.addAndChangePrivateKey(key, updateUI: true);
+    settingsProvider.addAndChangePrivateKey(key);
     nostr = await relayProvider.genNostr(signer);
 
     if (backAfterLogin && mounted) {